Anemoi/requirements.txt
# on ubuntu 22.04:
#
# apt install libglu1-mesa-dev
# npx playwright@1.53.1 install

aiohttp>=3.12.9
camel-ai[all]==0.2.70
docx2markdown>=0.1.1
httpx-sse>=0.4.0
logger>=1.4
loguru>=0.7.3
mcp==1.10.1
opencv-python>=4.11.0.86
python-dotenv>=1.1.1
retry>=0.9.2
starlette>=0.47.0
tqdm>=4.67.1
yarl>=1.20.0